Today's progress, fixing as many of the things commented out with #if joe_wip or with comment containing XXX:
* Renamed QueryableStatus to BackendStatus, and make some of the code,
particularly in the remote control stuff a little saner.

* Also renamed QueryableState to BackendState.

* Fix a bug in the LuceneQueryingDriver in which hits were being returned
twice.  This along with the URI remapping in the FSQ is what was causing
snippets to be broken.

* Fix up beagle-query to also show filter keyword mappings.

Still broken:

* Item counts are still reported wrong since they're reported by the
index count, not the source count.  This will probably require a Lucene
query, but we can do it entirely from the daemon and not call out to the
helper.

Still to come:

* Splitting up the singleton index into multiple indices.