Tool to index removable media. In the use cases that I envisioned, deciding when to index, what to index, how to identify the index, media insertion/removal notifications depend a lot on the users and the distribution that they are using. I have left these parts out of the tool.
Tool beagle-removable-index: Upon media insertion run
$ beagle-removable-index --mount-point <m> --media-name <n>

The tool is a wrapper to beagle-build-index and builds a _smart_ static
index. Trigger this from hal, manually or whatever media notification
mechanism you prefer.
Use the same command to update the index if there is some change in contents
of the media. Removable media is mostly used read-only or with seldom write
access. I dont feel they need not be constantly watched for changes.

Mount point: Mount point of the media. The paths to all the files and
dirctories in the subtree will be stored relative to the mount_point.

Media name: "Unique" name of the media. Use the hal ids or audio cd name or
whatever you want. This will be used to uniquely determine the actual static
index. The index will be created in ~/.beagle/Indexes/media_name

beagle-removable-index will store the mapping of "media-name <->
mount-point" in the settings. When returning search results, the relative
paths will be mapped to the actual paths based on the settings. Results
pointing to files/dirs not found on disk (i.e. unmounted media) will be
specially marked. Clients can read those markings and prompt user to insert
the required media (identified by name). Sophisticated handling of not-found
files can be dealt with later.