Merge r3413 and 3433-3434 onto the branch.
Add option --disable-textcache to beagled. This disallows caching data
for snippet purposes during this run. To prevent inconsistency, this
also deletes existing textcache for any document indexed in this
session. Usage of this option is highly discouraged and should be only
used on space-strapped machines.

Change --disable-textcache to --disable-text-cache to maintain
consistency with beagle-build-index's --enable-text-cache

svn path=/branches/beagle-0.2.16/; revision=3439