Merge from trunk:Util/
r3519: When serializing, wrap the buffer first in an UnclosableStream and then in a BufferedStream with an 8k buffer.  This gives us around a 10% speed up when serializing responses.

r3566: Have Beagle.Util.FileSystem.GetLastWriteTimeUtc() use the same rules as the File and Directory variants.  That is, return 1 Jan 1601 at midnight
UTC if the file doesn't exist rather than throwing an exception.  This
fixes bgo 417601.

r3671: Better GlobMatch routine.


svn path=/branches/beagle-0.2.16/; revision=3692