Merge r3857 from the trunk:
Fix a problem where processing large number of maildir files on the file
system would cause file descriptors to run out.  The glib-sharp binding
(at least up until 2.10) batches and unrefs disposed managed wrappers in
a glib timeout, which requires that a main loop be running.  While
beagled has one, beagle-build-index and beagle-extract-content do not.
So work around the problem by running "while
(GLib.MainContext.Pending()) GLib.MainContext.Iteration();" after
processing a batch of files, to ensure that wrappers are collected.
This fixes bgo #466891.

In the mail filter, log GMime initialization error messages better and
don't continuously retry to reinitialize.  If it's broken, it's going to
stay broken and we don't need to flood the logs with this.


svn path=/branches/beagle-0.2.16/; revision=3894