Simple RDF Executor that returns just the uri of the matches.
Use the included BeagleClient/rdf.cs to test.
$ gmcs -r:../Util/Util.dll -r:../BeagleClient/Beagle.dll rdf.cs
$ mono rdf.exe <query_string>

Only the uris will be printed.
For results from filesystem backend this will represent the internal uris. To get the actual uri, just to a usual query with the internal uri as the query string (e.g. beagle-query uid:eUcDXywxT0GJLSsejByK9A).

All of this uses the BeagleClient API, so can be suitably used in a single program to get only real uris for any query string.


svn path=/branches/beagle-rdf/; revision=4037